requirements.txt
streamlit
tensorflow
keras
scipy
numpy
opencv-python
Pillow